track print claim summary button clicks
