+++ src/claimApplication/claimantInfo/confirmation.js
@@ -1,3 +1,5 @@
+import { logEvent } from "../../modules/shared.mjs";
+
 export const confirmationAltTexts = [
   { id: 'ContentPlaceHolder1_ClaimantCertTab_TPConfirmation_imgbtnM01', alt: 'Submit Query' },
 ];
@@ -9,3 +11,12 @@ export const identifyingContent = {
   elementId: 'ContentPlaceHolder1_lblHeader',
   text: 'CLAIMANT CONFIRMATION',
 };
+
+export function trackPrintClaimSummaryButton() {
+  const printClaimSummaryButton = document.getElementById('ContentPlaceHolder1_ClaimantCertTab_TPConfirmation_btnContinue');
+  if (printClaimSummaryButton) {
+    printClaimSummaryButton.addEventListener('click', function() {
+      logEvent('Print Claim Summary Button Clicked', {});
+    });
+  }
+}
